requirements.txt
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
apache-beam==2.46.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==21.4.0
av==10.0.0
babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
braceexpand==0.1.7
cached-property==1.5.2
ccimport==0.4.2
cityscapesscripts==2.2.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
configargparse==1.5.3
contextlib2==21.6.0
crcmod==1.7
cumm-cu113==0.4.10
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
dill==0.3.1.1
docker-pycreds==0.4.0
docopt==0.6.2
dominate==2.7.0
dtcwt==0.12.0
efficientnet-pytorch==0.7.1
einops==0.6.0
entrypoints==0.4
exifread-nocycle==3.0.1
fairscale==0.4.6
fastavro==1.7.3
fasteners==0.18
fastjsonschema==2.16.3
filelock==3.9.0
fire==0.4.0
flask==2.2.5
flatbuffers==23.3.3
fonttools==4.38.0
fqdn==1.5.1
fsspec==2022.11.0
gast==0.2.2
gitdb==4.0.10
gitpython==3.1.31
glfw==2.5.9
google-api-core==2.11.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.7.0
google-cloud-bigquery-storage==2.16.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.12.0
google-cloud-language==1.3.2
google-cloud-pubsub==2.15.0
google-cloud-pubsublite==1.7.0
google-cloud-recommendations-ai==0.7.1
google-cloud-spanner==3.28.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.4.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.48.2
h5py==3.8.0
hdfs==2.7.0
httplib2==0.21.0
huggingface-hub==0.12.1
humanfriendly==10.0
hydra-core==1.3.2
imageio==2.26.0
imageio-ffmpeg==0.4.8
img2dataset==1.41.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
json-tricks==3.16.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-events==0.6.3
jupyter-packaging==0.12.3
jupyter-server==1.23.6
jupyter-server-fileid==0.9.0
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.22.1
jupyterlab-widgets==3.0.7
keras==2.11.0
keras-applications==1.0.8
kiwisolver==1.4.4
kornia==0.6.10
lark==1.1.5
libclang==15.0.6.1
lightning-utilities==0.8.0
lmdb==1.4.0
lpips==0.1.4
markdown-it-py==2.2.0
matlab==0.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdurl==0.1.2
medpy==0.4.0
mistune==2.0.5
mogrify==0.1
monai==1.1.0
multipledispatch==0.6.0
munch==3.0.0
mypy-extensions==1.0.0
natsort==8.3.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==4.0.2
ninja==1.11.1
nni==2.10
notebook==6.5.3
notebook-shim==0.2.2
nuscenes-devkit==1.1.10
oauth2client==4.1.3
objsize==0.6.1
omegaconf==2.2.3
open3d==0.13.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.8.7
overrides==6.5.0
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pathtools==0.1.2
pccm==0.4.7
pexpect==4.8.0
pickleshare==0.7.5
pip==23.1.2
piq==0.7.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
plotly==5.14.1
plyfile==0.8.1
portalocker==2.7.0
pretrainedmodels==0.7.4
prettytable==3.6.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
ptflops==0.6.9
ptyprocess==0.7.0
pyarrow==6.0.1
pybind11==2.10.4
pycocotools==2.0.6
pydot==1.4.2
pyglet==2.0.5
pyglfw==0.2.2
pygments==2.14.0
pymcubes==0.1.4
pymongo==3.13.0
pyopengl==3.1.6
pyparsing==3.0.9
pyquaternion==0.9.9
pyrr==0.10.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pythonwebhdfs==0.2.3
pytorch-lightning==1.9.5
pytz==2022.7.1
pywavelets==1.3.0
pyzmq==25.0.1
qtconsole==5.4.3
qtpy==2.3.1
qudida==0.0.4
regex==2022.10.31
responses==0.23.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
safetensors==0.3.1
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
segment-anything==1.0
segmentation-models-pytorch==0.3.3
send2trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.16.0
setproctitle==1.3.2
shapely==1.8.5
shortuuid==1.0.11
simpleitk==2.2.1
simplejson==3.18.4
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spconv-cu113==2.3.6
sqlparse==0.4.3
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-logger==0.1.0
tensorboardx==2.6
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.12.0
tensorflow-serving-api==2.11.0
terminado==0.17.1
tfx-bsl==1.12.0
thop==0.1.1-2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.2
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
torch-efficient-distloss==0.1.3
torch-ema==0.3
torch-geometric==2.3.0
torchmetrics==0.11.4
torchstat==0.0.7
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trimesh==3.21.5
typed-ast==1.5.4
typeguard==3.0.2
types-pyyaml==6.0.12.8
typing==3.7.4.3
ultralytics==8.0.98
uri-template==1.2.0
uritemplate==3.0.1
wandb==0.12.21
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
werkzeug==2.2.3
widgetsnbextension==4.0.7
y-py==0.5.9
yacs==0.1.8
yapf==0.32.0
ypy-websocket==0.8.2
zipp==3.15.0
zstandard==0.20.0